# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/communication/posts": {
      "get": {
        "operationId": "CommunicationController_posts",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "当前课程范围内讨论，分页、关键词、作者、创建时间范围筛选",
        "tags": ["交流"]
      },
      "post": {
        "operationId": "CommunicationController_createPost",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "课程发帖，校验课程与附件归属",
        "tags": ["交流"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "courseId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  "body": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 15000
                  },
                  "attachmentIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "maxItems": 10,
                    "default": []
                  },
                  "linkType": {
                    "type": "string",
                    "enum": ["chapter", "assignment", "question"]
                  },
                  "linkId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  }
                },
                "required": ["courseId", "title", "body"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/communication/posts/{id}": {
      "get": {
        "operationId": "CommunicationController_post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "讨论详情与分页回复；默认第一页20条",
        "tags": ["交流"]
      },
      "patch": {
        "operationId": "CommunicationController_updatePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "置顶、精华、关闭讨论；作者可标记解决",
        "tags": ["交流"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pinned": {
                    "type": "boolean"
                  },
                  "featured": {
                    "type": "boolean"
                  },
                  "closed": {
                    "type": "boolean"
                  },
                  "solved": {
                    "type": "boolean"
                  }
                },
                "required": [],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/communication/posts/{id}/replies": {
      "post": {
        "operationId": "CommunicationController_reply",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "回复或引用同一讨论中的回复",
        "tags": ["交流"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "body": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 15000
                  },
                  "quoteReplyId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "attachmentIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "maxItems": 10,
                    "default": []
                  }
                },
                "required": ["body"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/communication/contacts": {
      "get": {
        "operationId": "CommunicationController_contacts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "当前同课程的师生联系人，管理员不返回私人联系人",
        "tags": ["交流"]
      }
    },
    "/api/communication/classes": {
      "get": {
        "operationId": "CommunicationController_classes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "当前允许访问交流区的班级",
        "tags": ["交流"]
      }
    },
    "/api/communication/classes/{classId}/conversation": {
      "post": {
        "operationId": "CommunicationController_classConversation",
        "parameters": [
          {
            "name": "classId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "进入当前授权班级的交流会话",
        "tags": ["交流"]
      }
    },
    "/api/communication/conversations": {
      "get": {
        "operationId": "CommunicationController_conversations",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "会话列表、最后消息、准确未读数；重新校验当前成员资格",
        "tags": ["交流"]
      },
      "post": {
        "operationId": "CommunicationController_directConversation",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "建立或恢复同课程师生私人会话",
        "tags": ["交流"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  }
                },
                "required": ["userId"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/communication/conversations/{id}/messages": {
      "get": {
        "operationId": "CommunicationController_messages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "消息历史；afterId补取断线期间消息，稳定游标排序",
        "tags": ["交流"]
      },
      "post": {
        "operationId": "CommunicationController_send",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "幂等发送消息；相同clientId不同正文返回409",
        "tags": ["交流"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "body": {
                    "type": "string",
                    "maxLength": 15000,
                    "default": ""
                  },
                  "clientId": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 100,
                    "pattern": "^[a-zA-Z0-9_-]+$"
                  },
                  "attachmentIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "maxItems": 10,
                    "default": []
                  }
                },
                "required": ["clientId"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/communication/conversations/{id}/read": {
      "post": {
        "operationId": "CommunicationController_read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "推进已读位置，禁止游标回退",
        "tags": ["交流"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "messageId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  }
                },
                "required": ["messageId"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/communication/messages/{id}": {
      "delete": {
        "operationId": "CommunicationController_retract",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "两分钟内撤回本人消息",
        "tags": ["交流"]
      }
    },
    "/api/communication/reports": {
      "post": {
        "operationId": "CommunicationController_report",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "举报当前可访问的内容",
        "tags": ["交流"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "targetType": {
                    "type": "string",
                    "enum": ["post", "reply", "message"]
                  },
                  "targetId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 1000
                  }
                },
                "required": ["targetType", "targetId", "reason"],
                "additionalProperties": false
              }
            }
          }
        }
      },
      "get": {
        "operationId": "CommunicationController_reports",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "授权治理范围内举报元数据；不会暴露私人会话正文",
        "tags": ["交流"]
      }
    },
...
    "/api/communication/blocks/{userId}": {
      "delete": {
        "operationId": "CommunicationController_unblock",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "解除本人设置的屏蔽",
        "tags": ["交流"]
      }
    },
...
  },
  "info": {
    "title": "知学 · 学习管理系统 API",
    "description": "同源 Cookie 会话。写请求必须携带登录响应中的 x-csrf-token。所有教学资源逐请求执行机构、当前角色、课程成员资格和对象归属校验。分页 page/pageSize，错误 {error:{code,message,fields,requestId}}。详细参数及示例见 docs/*-api.md。",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lms_session"
      }
    },
...
  }
}